Disable autoDiscover
When you don't have a form, the autoDiscover feature of DropZone creates a new instance of DropZone with option.url=null. This results in an Error thrown in dropzone.js on line 595, before the instantiation on line 68 of DropZone.php is executed. 
I have added Dropzone.autoDiscover = false; and moved the DropZoneAsset::register($view) down to make sure that disabling autoDiscover is being applied.